+++ tripService/src/main/java/ru/itmo/tripService/kafka/model/CarControlMessage.java
@@ -9,8 +9,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CarControlMessage {
 
-    private Integer tripId;
-    private Integer carId;
+    private Long tripId;
+    private Long carId;
 
     private Double latitude;
     private Double longitude;
